+++ src/pages/colinks/ActivityPage.tsx
@@ -6,8 +6,8 @@ import { Helmet } from 'react-helmet';
 import { LoadingIndicator } from '../../components/LoadingIndicator';
 import { ActivityList } from '../../features/activities/ActivityList';
 import { CoLinksContext } from '../../features/colinks/CoLinksContext';
-import { PROMPTS } from '../../features/colinks/PostForm';
-import { PostForm2 } from '../../features/colinks/PostForm2';
+import { ContributionForm2 } from '../../features/colinks/ContributionForm2';
+import { PROMPTS } from '../../features/colinks/prompts';
 import { RecentCoLinkTransactions } from '../../features/colinks/RecentCoLinkTransactions';
 import { RightColumnSection } from '../../features/colinks/RightColumnSection';
 import { SimilarProfiles } from '../../features/colinks/SimilarProfiles';
@@ -81,11 +81,14 @@ const CoLinksActivityPageContents = ({
             }}
           >
             {targetBalance !== undefined && targetBalance > 0 && (
-              <PostForm2
+              <ContributionForm2
                 label={
                   <Text size={'medium'} semibold color={'heading'}>
                     {currentPrompt(promptOffset)}
                   </Text>
+                }
+                placeholder={
+                  'Take inspiration from the prompt, or post whatever you want'
                 }
                 showLoading={showLoading}
                 onSave={() => setShowLoading(true)}
