+++ src/pages/colinks/NotificationsPage.tsx
@@ -23,9 +23,11 @@ import { CSS } from '../../stitches.config';
 import {
   AppLink,
   Avatar,
+  Box,
   ContentHeader,
   Flex,
   Link,
+  MarkdownPreview,
   Panel,
   Text,
 } from '../../ui';
@@ -307,14 +309,13 @@ export const MentionReply = ({
               </Text>
             </Flex>
           </Flex>
-          <Text
-            color={'default'}
+          <Box
             as={NavLink}
             to={coLinksPaths.post(`${reply.activity_id}`)}
             css={{ textDecoration: 'none' }}
           >
-            {reply.reply}
-          </Text>
+            <MarkdownPreview display source={reply.reply}></MarkdownPreview>
+          </Box>
         </Flex>
       </Flex>
     </NotificationItem>
@@ -435,14 +436,16 @@ export const MentionPost = ({
               </Text>
             </Flex>
           </Flex>
-          <Text
-            color={'default'}
+          <Box
             as={NavLink}
             to={coLinksPaths.post(`${post?.activity?.id}`)}
             css={{ textDecoration: 'none' }}
           >
-            {post.description}
-          </Text>
+            <MarkdownPreview
+              display
+              source={post.description}
+            ></MarkdownPreview>
+          </Box>
         </Flex>
       </Flex>
     </NotificationItem>
